popovers: Move left sidebar popovers to a subdirectory.
