+++ web/src/topic_popover.js
@@ -2,7 +2,7 @@ import ClipboardJS from "clipboard";
 import $ from "jquery";
 
 import render_delete_topic_modal from "../templates/confirm_dialog/confirm_delete_topic.hbs";
-import render_left_sidebar_topic_actions_popover from "../templates/popovers/left_sidebar_topic_actions_popover.hbs";
+import render_left_sidebar_topic_actions_popover from "../templates/popovers/left_sidebar/left_sidebar_topic_actions_popover.hbs";
 
 import * as confirm_dialog from "./confirm_dialog";
 import {$t_html} from "./i18n";
