+++ web/templates/popovers/left_sidebar/left_sidebar_all_messages_popover.hbs
