+++ web/templates/popovers/left_sidebar/left_sidebar_condensed_views_popover.hbs
