+++ web/templates/popovers/left_sidebar/left_sidebar_drafts_popover.hbs
