+++ web/templates/popovers/left_sidebar/left_sidebar_inbox_popover.hbs
