+++ web/templates/popovers/left_sidebar/left_sidebar_recent_view_popover.hbs
