+++ web/templates/popovers/left_sidebar/left_sidebar_stream_actions_popover.hbs
@@ -5,7 +5,7 @@
             <ul class="popover-menu-inner-list">
                 <li class="popover-stream-header text-item popover-menu-inner-list-item">
                     <span class="stream-privacy-original-color-{{stream.stream_id}} stream-privacy filter-icon" style="color: {{stream.color}}">
-                        {{> ../stream_privacy
+                        {{> ../../stream_privacy
                           invite_only=stream.invite_only
                           is_web_public=stream.is_web_public }}
                     </span>
